Include points in model
The previous model only used games and sets. This assigns a value to
every point (0, 15, 30, 40, AD, plus a tiebreak). The delta between the
players is used to calibrate the model.

The result is much better accuracy even with fewer samples. It also
results in the final probability approaching 100% in many cases (rather
than 70-85% previously).